requirements.txt
mysql-connector-python==9.1.0
kafka-python==2.0.2
pyspark==3.5.3
python-dotenv==1.0.1